+++ src/test-mvnorm.cpp
@@ -1,5 +1,5 @@
 #include <testthat.h>
-#include "mvnorm.h"
+#include "internal.h"
 
 context("dmvnorm_cpp") {
 
@@ -16,20 +16,6 @@ context("dmvnorm_cpp") {
     expect_true(density_log_rounded == -0.919);
   }
 
-  test_that("univariate normal density can be computed (degenerate case)") {
-    arma::vec x_1 = arma::vec(1).fill(1);
-    arma::vec x_2 = arma::vec(1).fill(1.1);
-    arma::mat Sigma = arma::mat(1, 1, arma::fill::zeros);
-    double density_1 = dmvnorm_cpp(x_1, x_1, Sigma, false);
-    expect_true(std::isinf(density_1));
-    double density_1_log = dmvnorm_cpp(x_1, x_1, Sigma, true);
-    expect_true((std::isinf(density_1_log) && density_1_log > 0));
-    double density_2 = dmvnorm_cpp(x_1, x_2, Sigma, false);
-    expect_true(density_2 == 0.0);
-    double density_2_log = dmvnorm_cpp(x_1, x_2, Sigma, true);
-    expect_true((std::isinf(density_2_log) && density_2_log < 0));
-  }
-
   test_that("multivariate normal density can be computed") {
     arma::vec x = arma::zeros<arma::vec>(2);
     arma::vec mean = arma::zeros<arma::vec>(2);
@@ -43,18 +29,24 @@ context("dmvnorm_cpp") {
     expect_true(density_log_rounded == -1.838);
   }
 
-  test_that("multivariate normal density can be computed (degenerate case)") {
-    arma::vec x_1 = arma::vec(2).fill(1);
-    arma::vec x_2 = arma::vec(2).fill(1.1);
-    arma::mat Sigma = arma::mat(2, 2, arma::fill::zeros);
-    double density_1 = dmvnorm_cpp(x_1, x_1, Sigma, false);
-    expect_true(std::isinf(density_1));
-    double density_1_log = dmvnorm_cpp(x_1, x_1, Sigma, true);
-    expect_true((std::isinf(density_1_log) && density_1_log > 0));
-    double density_2 = dmvnorm_cpp(x_1, x_2, Sigma, false);
-    expect_true(density_2 == 0.0);
-    double density_2_log = dmvnorm_cpp(x_1, x_2, Sigma, true);
-    expect_true((std::isinf(density_2_log) && density_2_log < 0));
+}
+
+context("pmvnorm_cpp") {
+
+  test_that("univariate normal CDF can be computed") {
+    arma::vec x = arma::zeros<arma::vec>(1);
+    arma::vec mean = arma::zeros<arma::vec>(1);
+    arma::mat Sigma = arma::eye<arma::mat>(1,1);
+    double prob = pmvnorm_cpp(x, mean, Sigma);
+    expect_true(prob == 0.5);
+  }
+
+  test_that("multivariate normal CDF can be computed") {
+    arma::vec x = arma::zeros<arma::vec>(2);
+    arma::vec mean = arma::zeros<arma::vec>(2);
+    arma::mat Sigma = arma::eye<arma::mat>(2,2);
+    double prob = pmvnorm_cpp(x, mean, Sigma);
+    expect_true(prob == 0.25);
   }
 
 }