+++ src/wishart.cpp
@@ -1,10 +1,10 @@
 // [[Rcpp::depends("RcppArmadillo")]]
 #include <RcppArmadillo.h>
 
-//' @rdname dwishart
-//' @export
+// [[Rcpp::interfaces(cpp)]]
 // [[Rcpp::export]]
-double dwishart_cpp(
+
+double dwishart(
   arma::mat const& x, int const& df, arma::mat const& scale,
   bool log = false, bool inv = false
 ) {
@@ -31,28 +31,32 @@ double dwishart_cpp(
   return density;
 }
 
-//' @rdname dwishart
-//' @export
+// [[Rcpp::interfaces(cpp)]]
 // [[Rcpp::export]]
-arma::mat rwishart_cpp(
+
+arma::mat rwishart(
   double df, arma::mat const& scale, bool inv = false
 ) {
   int m = scale.n_rows;
-  arma::mat T = arma::zeros<arma::mat>(m,m);
-  for(int i = 0; i < m; i++) {
-   T(i,i) = std::sqrt(R::rchisq(df-i));
+  arma::mat T = arma::zeros<arma::mat>(m, m);
+  for (int i = 0; i < m; ++i) {
+    T(i, i) = std::sqrt(R::rchisq(df - i));
+    for (int j = 0; j < i; ++j) {
+      T(i, j) = R::rnorm(0.0, 1.0);
+    }
   }
-  for(int j = 0; j < m; j++) {
-   for(int i = j+1; i < m; i++) {
-     T(i,j) = R::rnorm(0.0, 1.0);
-   }
-  }
-  arma::mat C = trans(T)*chol(scale);
-  arma::mat CI = solve(trimatu(C), arma::eye(m,m));
+  arma::mat L;
   if (inv) {
-   return trans(C) * C;
+    arma::mat scale_inv = arma::inv_sympd(scale);
+    L = arma::chol(scale_inv, "lower");
   } else {
-   return CI * trans(CI);
+    L = arma::chol(scale, "lower");
+  }
+  arma::mat C = L * T;
+  arma::mat A = C * C.t();
+  if (inv) {
+    return A.i();
+  } else {
+    return A;
   }
 }
-