+++ NEWS.md
@@ -1,3 +1,19 @@
+# oeli 0.7.5
+
+* Moved header files from `../src/` to `../inst/include/` to make them available externally.
+
+* Removed degenerate case from `dmvnorm()`.
+
+* Fixed a bug in sampling from the inverse Wishart distribution.
+
+* Added function `pmvnorm()` that calculates the (multivariate) Gaussian CDF.
+
+* Added function `gaussian_tv()` that calculates Gaussian total variation.
+
+* Added function `equidistant_vectors()` that generates equidistant vectors in Euclidean space.
+
+* Added functions `dmixnorm()`, `pmixnorm()`, and `rmixnorm()` that compute the density, cumulative distribution function, and samples from a mixture of multivariate normal distribution.
+
 # oeli 0.7.4
 
 * Added R6 object `Simulator` that provides some convenience when running simulations.